+++ packages/react/src/components/Accordion/Accordion.tsx
@@ -6,26 +6,27 @@ import classes from './Accordion.module.css';
 
 export type AccordionProps = {
   /** Accordion background color */
-  color?: 'primary' | 'secondary' | 'tertiary' | 'neutral' | 'subtle';
+  color?: 'first' | 'second' | 'third' | 'neutral' | 'subtle';
   /** Show border */
   border?: boolean;
   /** Instances of `Accordion.Item` */
   children: React.ReactNode;
 } & HTMLAttributes<HTMLDivElement>;
 
-export const Accordion = forwardRef<HTMLDivElement, AccordionProps>(
-  ({ border = false, color = 'neutral', className, ...rest }, ref) => (
-    <div
-      {...rest}
-      className={cn(
-        classes.accordion,
-        classes[color],
-        {
-          [classes.border]: border,
-        },
-        className,
-      )}
-      ref={ref}
-    />
-  ),
-);
+export const Accordion = forwardRef<
+  HTMLDivElement,
+  AccordionProps & { children: React.ReactNode }
+>(({ border = false, color = 'neutral', className, ...rest }, ref) => (
+  <div
+    {...rest}
+    className={cn(
+      classes.accordion,
+      classes[color],
+      {
+        [classes.border]: border,
+      },
+      className,
+    )}
+    ref={ref}
+  />
+));