+++ packages/react/src/components/ToggleButtonGroup/index.ts
@@ -1,5 +1,0 @@
-export { ToggleButtonGroup } from './ToggleButtonGroup';
-export type {
-  ToggleButtonGroupProps,
-  ToggleButtonProps,
-} from './ToggleButtonGroup';